+++ src/Async/Async.cpp
@@ -2,9 +2,10 @@
 #include "Async/Executor.hpp"
 #include "MultiThread/AutoResetEvent.hpp"
 #include <boost/smart_ptr/intrusive_ptr.hpp>
+#include <cstddef>
 #include <mutex>
 
-AsyncMultiLoop::AsyncMultiLoop(int iterations,
+AsyncMultiLoop::AsyncMultiLoop(size_t iterations,
                                std::function<void(size_t)> iterLambda)
     : _iterLambda(iterLambda), _iterations(iterations) {
   Scheduler::main().schedule(boost::intrusive_ptr<Task>{this});
